src/agent/quant_research_agent.py
import os
import git
import markdown
from datetime import datetime
import torch
import yfinance as yf
import pandas as pd
from pathlib import Path

class QuantResearchAgent:
    def __init__(self, repo_path):
        self.repo_path = Path(repo_path)
        self.research_path = self.repo_path / "research"
        self.models_path = self.repo_path / "models"
        self.data_path = self.repo_path / "data"
        
        # Create directories if they don't exist
        for path in [self.research_path, self.models_path, self.data_path]:
            path.mkdir(parents=True, exist_ok=True)
        
        # Initialize git repo
        self.repo = git.Repo.init(repo_path)

    def create_research_document(self, title, content):
        """Create a new research document in markdown format"""
        filename = f"{datetime.now().strftime('%Y%m%d')}_{title.lower().replace(' ', '_')}.md"
        filepath = self.research_path / filename
        
        with open(filepath, 'w') as f:
            f.write(content)
        
        return filepath

    def save_model(self, model, model_name):
        """Save a PyTorch model"""
        model_path = self.models_path / f"{model_name}.pth"
        torch.save(model.state_dict(), model_path)
        return model_path

    def commit_changes(self, message):
        """Commit changes to the repository"""
        self.repo.index.add('*')
        self.repo.index.commit(message)

    def fetch_market_data(self, symbol, start_date, end_date):
        """Fetch market data using yfinance"""
        data = yf.download(symbol, start=start_date, end=end_date)
        return data

    def analyze_strategy(self, data, strategy_name):
        """Analyze a trading strategy and create a research document"""
        # Implement strategy analysis logic here
        pass

    def train_model(self, data, model_type):
        """Train a machine learning model"""
        # Implement model training logic here
        pass
