requirements.txt
numpy==1.24.3
pandas==2.0.2
scikit-learn==1.2.2
yfinance==0.2.18
pytest==7.3.1
python-dotenv==1.0.0
gitpython==3.1.31
markdown==3.4.3
torch==2.0.1
matplotlib==3.7.1
